Move definitions specific to the concrete evaluator in to separate modules.  The module structure needs to be a bit more cut up than I would like to avoid module import cycles.
