+++ cryptol.cabal
@@ -153,6 +153,7 @@ library
                        Cryptol.Eval,
                        Cryptol.Eval.Arch,
                        Cryptol.Eval.Concrete,
+                       Cryptol.Eval.Concrete.Value,
                        Cryptol.Eval.Env,
                        Cryptol.Eval.Generic,
                        Cryptol.Eval.Monad,
@@ -161,7 +162,6 @@ library
                        Cryptol.Eval.Type,
                        Cryptol.Eval.Value,
 
-                       Cryptol.Testing.Concrete,
                        Cryptol.Testing.Random,
 
                        Cryptol.Symbolic,
