+++ src/Cryptol/Testing/Random.hs
@@ -9,24 +9,33 @@
 -- This module generates random values for Cryptol types.
 
 {-# LANGUAGE BangPatterns #-}
+{-# LANGUAGE RecordWildCards #-}
 {-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE Trustworthy #-}
 module Cryptol.Testing.Random where
 
-import Cryptol.Eval.Monad     (ready,runEval,EvalOpts,io)
-import Cryptol.Eval.Value     (Value,GenValue(..),SeqMap(..), WordValue(..), BitWord(..))
-import qualified Cryptol.Testing.Concrete as Conc
+import qualified Control.Exception as X
+import Control.Monad          (join, liftM2)
+import Data.Bits              ( (.&.), shiftR )
+import Data.List              (unfoldr, genericTake, genericIndex, genericReplicate)
+import Data.Map (Map)
+import qualified Data.Map as Map
+import qualified Data.Sequence as Seq
+
+import System.Random          (RandomGen, split, random, randomR)
+import System.Random.TF.Gen   (seedTFGen)
+
+import Cryptol.Eval.Concrete.Value
+import Cryptol.Eval.Monad     (ready,runEval,EvalOpts,io,Eval,EvalError(..))
+import Cryptol.Eval.Type      (TValue(..), tValTy)
+import Cryptol.Eval.Value     (GenValue(..),SeqMap(..), WordValue(..), BitWord(..),
+                               ppValue, defaultPPOpts, finiteSeqMap)
+import Cryptol.Eval.Generic   (zeroV)
 import Cryptol.TypeCheck.AST  (Type(..), TCon(..), TC(..), tNoUser, tIsFun)
 import Cryptol.TypeCheck.SimpType(tRebuild')
 
 import Cryptol.Utils.Ident    (Ident)
 import Cryptol.Utils.Panic    (panic)
-
-import Control.Monad          (join)
-import Data.List              (unfoldr, genericTake, genericIndex)
-import System.Random          (RandomGen, split, random, randomR)
-import Data.Map (Map)
-import qualified Data.Map as Map
-import qualified Data.Sequence as Seq
 
 type Gen g x = Integer -> g -> (IO (GenValue x), g)
 
@@ -44,12 +53,12 @@ runOneTest :: RandomGen g
         -> [Gen g ()] -- ^ Argument generators
         -> Integer -- ^ Size
         -> g
-        -> IO (Conc.TestResult, g)
+        -> IO (TestResult, g)
 runOneTest evOpts fun argGens sz g0 = do
   let (args, g1) = foldr mkArg ([], g0) argGens
       mkArg argGen (as, g) = let (a, g') = argGen sz g in (a:as, g')
   args' <- sequence args
-  result <- Conc.runOneTest evOpts fun args'
+  result <- evalTest evOpts fun args'
   return (result, g1)
 
 returnOneTest :: RandomGen g
@@ -91,14 +100,14 @@ returnTests g evo gens fun num = go gens g 0
            return ((inputs, output) : more)
 
 {- | Given a (function) type, compute generators for the function's
-arguments. This is like 'testableType', but allows the result to be
+arguments. This is like 'testableTypeGenerators', but allows the result to be
 any finite type instead of just @Bit@. -}
 dumpableType :: forall g. RandomGen g => Type -> Maybe [Gen g ()]
 dumpableType ty =
   case tIsFun ty of
     Just (t1, t2) ->
       do g  <- randomValue () t1
-         as <- testableType t2
+         as <- testableTypeGenerators t2
          return (g : as)
     Nothing ->
       do (_ :: Gen g ()) <- randomValue () ty
@@ -107,12 +116,12 @@ dumpableType ty =
 {- | Given a (function) type, compute generators for
 the function's arguments. Currently we do not support polymorphic functions.
 In principle, we could apply these to random types, and test the results. -}
-testableType :: RandomGen g => Type -> Maybe [Gen g ()]
-testableType ty =
+testableTypeGenerators :: RandomGen g => Type -> Maybe [Gen g ()]
+testableTypeGenerators ty =
   case tNoUser ty of
     TCon (TC TCFun) [t1,t2] ->
       do g  <- randomValue () t1
-         as <- testableType t2
+         as <- testableTypeGenerators t2
          return (g : as)
     TCon (TC TCBit) [] -> return []
     _ -> Nothing
@@ -221,3 +230,184 @@ randomRecord gens sz g0 =
     mk g gen =
       let (v, g') = gen sz g
       in seq v (g', io v)
+
+
+-- Random Values ---------------------------------------------------------------
+
+-- | Produce a random value with the given seed. If we do not support
+-- making values of the given type, return zero of that type.
+-- TODO: do better than returning zero
+randomV :: BitWord sym => sym -> TValue -> Integer -> Eval (GenValue sym)
+randomV sym ty seed =
+  case randomValue sym (tValTy ty) of
+    Nothing -> zeroV sym ty
+    Just gen ->
+      -- unpack the seed into four Word64s
+      let mask64 = 0xFFFFFFFFFFFFFFFF
+          unpack s = fromInteger (s .&. mask64) : unpack (s `shiftR` 64)
+          [a, b, c, d] = take 4 (unpack seed)
+      in io $ fst $ gen 100 $ seedTFGen (a, b, c, d)
+
+
+-- | A test result is either a pass, a failure due to evaluating to
+-- @False@, or a failure due to an exception raised during evaluation
+data TestResult
+  = Pass
+  | FailFalse [Value]
+  | FailError EvalError [Value]
+
+isPass :: TestResult -> Bool
+isPass Pass = True
+isPass _    = False
+
+-- | Apply a testable value to some arguments.
+-- Note that this function assumes that the values come from a call to
+-- `testableType` (i.e., things are type-correct). We run in the IO
+-- monad in order to catch any @EvalError@s.
+evalTest :: EvalOpts -> Value -> [Value] -> IO TestResult
+evalTest evOpts v0 vs0 = run `X.catch` handle
+  where
+    run = do
+      result <- runEval evOpts (go v0 vs0)
+      if result
+        then return Pass
+        else return (FailFalse vs0)
+    handle e = return (FailError e vs0)
+
+    go :: Value -> [Value] -> Eval Bool
+    go (VFun f) (v : vs) = join (go <$> (f (ready v)) <*> return vs)
+    go (VFun _) []       = panic "Not enough arguments while applying function"
+                           []
+    go (VBit b) []       = return b
+    go v vs              = do vdoc    <- ppValue () defaultPPOpts v
+                              vsdocs  <- mapM (ppValue () defaultPPOpts) vs
+                              panic "Type error while running test" $
+                               [ "Function:"
+                               , show vdoc
+                               , "Arguments:"
+                               ] ++ map show vsdocs
+
+{- | Given a (function) type, compute all possible inputs for it.
+We also return the types of the arguments and
+the total number of test (i.e., the length of the outer list. -}
+testableType :: Type -> Maybe (Maybe Integer, [Type], [[Value]])
+testableType ty =
+  case tNoUser ty of
+    TCon (TC TCFun) [t1,t2] ->
+      do let sz = typeSize t1
+         (tot,ts,vss) <- testableType t2
+         return (liftM2 (*) sz tot, t1:ts, [ v : vs | v <- typeValues t1, vs <- vss ])
+    TCon (TC TCBit) [] -> return (Just 1, [], [[]])
+    _ -> Nothing
+
+{- | Given a fully-evaluated type, try to compute the number of values in it.
+Returns `Nothing` for infinite types, user-defined types, polymorphic types,
+and, currently, function spaces.  Of course, we can easily compute the
+sizes of function spaces, but we can't easily enumerate their inhabitants. -}
+typeSize :: Type -> Maybe Integer
+typeSize ty =
+  case ty of
+    TVar _      -> Nothing
+    TUser _ _ t -> typeSize t
+    TRec fs     -> product <$> mapM (typeSize . snd) fs
+    TCon (TC tc) ts ->
+      case (tc, ts) of
+        (TCNum _, _)     -> Nothing
+        (TCInf, _)       -> Nothing
+        (TCBit, _)       -> Just 2
+        (TCInteger, _)   -> Nothing
+        (TCIntMod, [sz]) -> case tNoUser sz of
+                              TCon (TC (TCNum n)) _ -> Just n
+                              _                     -> Nothing
+        (TCIntMod, _)    -> Nothing
+        (TCSeq, [sz,el]) -> case tNoUser sz of
+                              TCon (TC (TCNum n)) _ -> (^ n) <$> typeSize el
+                              _                     -> Nothing
+        (TCSeq, _)       -> Nothing
+        (TCFun, _)       -> Nothing
+        (TCTuple _, els) -> product <$> mapM typeSize els
+        (TCAbstract _, _) -> Nothing
+        (TCNewtype _, _) -> Nothing
+
+    TCon _ _ -> Nothing
+
+
+{- | Returns all the values in a type.  Returns an empty list of values,
+for types where 'typeSize' returned 'Nothing'. -}
+typeValues :: Type -> [Value]
+typeValues ty =
+  case ty of
+    TVar _      -> []
+    TUser _ _ t -> typeValues t
+    TRec fs     -> [ VRecord (fmap ready xs)
+                   | xs <- sequence (fmap typeValues (Map.fromList fs))
+                   ]
+    TCon (TC tc) ts ->
+      case tc of
+        TCNum _     -> []
+        TCInf       -> []
+        TCBit       -> [ VBit False, VBit True ]
+        TCInteger   -> []
+        TCIntMod    ->
+          case map tNoUser ts of
+            [ TCon (TC (TCNum n)) _ ] | 0 < n ->
+              [ VInteger x | x <- [ 0 .. n - 1 ] ]
+            _ -> []
+        TCSeq       ->
+          case map tNoUser ts of
+            [ TCon (TC (TCNum n)) _, TCon (TC TCBit) [] ] ->
+              [ VWord n (ready (WordVal (BV n x))) | x <- [ 0 .. 2^n - 1 ] ]
+
+            [ TCon (TC (TCNum n)) _, t ] ->
+              [ VSeq n (finiteSeqMap (map ready xs))
+              | xs <- sequence $ genericReplicate n
+                               $ typeValues t ]
+            _ -> []
+
+
+        TCFun       -> []  -- We don't generate function values.
+        TCTuple _   -> [ VTuple (map ready xs)
+                       | xs <- sequence (map typeValues ts)
+                       ]
+        TCAbstract _ -> []
+        TCNewtype _ -> []
+
+    TCon _ _ -> []
+
+--------------------------------------------------------------------------------
+-- Driver function
+
+data TestSpec m s = TestSpec {
+    testFn :: Integer -> s -> m (TestResult, s)
+  , testProp :: String -- ^ The property as entered by the user
+  , testTotal :: Integer
+  , testPossible :: Maybe Integer -- ^ Nothing indicates infinity
+  , testRptProgress :: Integer -> Integer -> m ()
+  , testClrProgress :: m ()
+  , testRptFailure :: TestResult -> m ()
+  , testRptSuccess :: m ()
+  }
+
+data TestReport = TestReport {
+    reportResult :: TestResult
+  , reportProp :: String -- ^ The property as entered by the user
+  , reportTestsRun :: Integer
+  , reportTestsPossible :: Maybe Integer
+  }
+
+runTests :: Monad m => TestSpec m s -> s -> m TestReport
+runTests TestSpec {..} st0 = go 0 st0
+  where
+  go testNum _ | testNum >= testTotal = do
+    testRptSuccess
+    return $ TestReport Pass testProp testNum testPossible
+  go testNum st =
+   do testRptProgress testNum testTotal
+      res <- testFn (div (100 * (1 + testNum)) testTotal) st
+      testClrProgress
+      case res of
+        (Pass, st') -> do -- delProgress -- unnecessary?
+          go (testNum + 1) st'
+        (failure, _st') -> do
+          testRptFailure failure
+          return $ TestReport failure testProp testNum testPossible
